+++ SpatialGDK/Source/SpatialGDK/Private/Utils/ComponentReader.cpp
@@ -168,7 +168,38 @@ void ComponentReader::ApplySchemaObject(Schema_Object* ComponentObject, UObject&
 	bool bIsClient = NetDriver->GetNetMode() == NM_Client;
 	bool bEventTracerEnabled = EventTracer != nullptr;
 
-	FSpatialConditionMapFilter ConditionMap(&Channel, bIsClient);
+	// RemoteRole (which is swapped into the local Actor's Role property) and bRepPhysics are used to construct the condition map.
+	// We must check to see if these properties have just been received, before we construct the condition map for receiving everything
+	// else.
+	ENetRole ActorRole = Channel.Actor->Role;
+	bool bRepPhysics = Channel.Actor->GetReplicatedMovement().bRepPhysics;
+	if (EnumHasAnyFlags(Replicator->RepLayout->GetFlags(), ERepLayoutFlags::IsActor))
+	{
+		for (uint32 FieldId : UpdatedIds)
+		{
+			int32 CmdIndex = BaseHandleToCmdIndex[FieldId - 1].CmdIndex;
+			const FRepLayoutCmd& Cmd = Cmds[CmdIndex];
+			if (UNLIKELY((int32)AActor::ENetFields_Private::RemoteRole == Cmd.ParentIndex))
+			{
+				ActorRole = (ENetRole)Schema_IndexUint32(ComponentObject, FieldId, 0);
+				// Downgrade role from AutonomousProxy to SimulatedProxy if we aren't authoritative over
+				// the client RPCs component. A race condition can occur here, if we check out the actor
+				// before the channel is Authoritative of it. The race gets handled in
+				// ActorSystem::HandleActorAuthority where we check USpatialActorChannel::IsAutonomousProxyOnAuthority.
+				if (ActorRole == ROLE_AutonomousProxy && !bIsAuthServer && !bAutonomousProxy)
+				{
+					ActorRole = ROLE_SimulatedProxy;
+				}
+			}
+			else if (UNLIKELY((int32)AActor::ENetFields_Private::ReplicatedMovement == Cmd.ParentIndex))
+			{
+				// Pull out bRepPhysics, based on FRepMovement::NetSerialize
+				TArray<uint8> ValueData = IndexBytesFromSchema(ComponentObject, FieldId, 0);
+				bRepPhysics = ValueData[0] & (1 << 1) ? true : false;
+			}
+		}
+	}
+	FSpatialConditionMapFilter ConditionMap(&Channel, bIsClient, ActorRole, bRepPhysics);
 
 	{
 		// Scoped to exclude OnRep callbacks which are already tracked per OnRep function
@@ -304,6 +335,7 @@ void ComponentReader::ApplySchemaObject(Schema_Object* ComponentObject, UObject&
 					// Downgrade role from AutonomousProxy to SimulatedProxy if we aren't authoritative over
 					// the client RPCs component.
 					GDK_PROPERTY(ByteProperty)* ByteProperty = GDK_CASTFIELD<GDK_PROPERTY(ByteProperty)>(Cmd.Property);
+					Channel.SetAutonomousProxyOnAuthority(ByteProperty->GetPropertyValue(Data) == ROLE_AutonomousProxy);
 					if (!bIsAuthServer && !bAutonomousProxy && ByteProperty->GetPropertyValue(Data) == ROLE_AutonomousProxy)
 					{
 						ByteProperty->SetPropertyValue(Data, ROLE_SimulatedProxy);