+++ SpatialGDK/Source/SpatialGDKFunctionalTests/Public/SpatialFunctionalTestFlowController.h
@@ -24,9 +24,8 @@ public:
 
 	void GetLifetimeReplicatedProps(TArray<FLifetimeProperty>& OutLifetimeProps) const override;
 
-	virtual void BeginPlay() override;
-
-	virtual void OnAuthorityGained() override;
+	virtual void OnActorReady(bool bHasAuthority) override;
+	virtual void OnClientOwnershipGained() override;
 
 	virtual void Tick(float DeltaSeconds) override;
 
@@ -79,14 +78,13 @@ public:
 	bool HasAckFinishedTest() const { return bHasAckFinishedTest; }
 
 	UFUNCTION()
+	void TrySetReadyToRunTest();
+	UFUNCTION()
 	void DeregisterFlowController();
 
 private:
 	// Current Step being executed
 	SpatialFunctionalTestStep CurrentStep;
-
-	UPROPERTY(ReplicatedUsing = OnReadyToRegisterWithTest)
-	uint8 bReadyToRegisterWithTest : 1;
 
 	UPROPERTY(Replicated)
 	bool bIsReadyToRunTest;
@@ -95,12 +93,7 @@ private:
 	bool bHasAckFinishedTest;
 
 	UFUNCTION()
-	void OnReadyToRegisterWithTest();
-
-	UFUNCTION()
 	void OnRep_OwningTest();
-
-	void TryRegisterFlowControllerWithOwningTest();
 
 	UFUNCTION(Server, Reliable)
 	void ServerSetReadyToRunTest(bool bIsReady);